# openapi/payops-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayOps Merchant API",
    "version": "0.3.0",
    "description": "Organization-scoped Solana stablecoin invoicing with hosted exact-amount Solana Pay checkout and finalized payment status."
  },
  "servers": [{ "url": "https://api.example.com" }],
  "security": [{ "sessionCookie": [] }, { "apiKey": [] }],
  "paths": {
...
    "/v1/auth/bootstrap/accept": {
      "post": {
        "operationId": "acceptBootstrapInvitation",
        "security": [],
        "description": "Accept a single-use operator-issued bootstrap invitation. A trusted Origin header is required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["token", "email", "name", "password"],
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 43,
                    "maxLength": 43,
                    "pattern": "^[A-Za-z0-9_-]{43}$"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 128
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bootstrap owner created; email verification is still required",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["userId", "organizationId", "email"],
                  "properties": {
                    "userId": { "$ref": "#/components/schemas/Uuid" },
                    "organizationId": { "$ref": "#/components/schemas/Uuid" },
                    "email": { "type": "string", "format": "email" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/organization": {
      "get": {
        "operationId": "getOrganization",
        "responses": {
          "200": {
            "description": "Active organization and effective permissions",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationContext" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "payops.session_token"
      },
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "headers": {
      "RequestId": {
        "description": "Canonical request UUID",
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
    },
...
    "responses": {
...
      "ApiError": {
        "description": "Stable API error",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
...
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": { "type": "string", "minLength": 1, "maxLength": 128 },
          "message": { "type": "string", "minLength": 1, "maxLength": 256 },
          "requestId": { "$ref": "#/components/schemas/Uuid" },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": ["string", "number", "boolean", "null"]
            }
          }
        }
      },
...
      "OrganizationContext": {
        "type": "object",
        "additionalProperties": false,
        "required": ["organizationId", "actorKind", "permissions"],
        "properties": {
          "organizationId": { "$ref": "#/components/schemas/Uuid" },
          "actorKind": { "type": "string", "enum": ["session", "api_key"] },
          "permissions": {
            "type": "object",
            "additionalProperties": { "type": "boolean" }
          }
        }
      },
...
    }
  }
}